Add measure to T/t and V/v examples.
The voice against which a note is dissonant must have one more note in order for certain dissonances to be detected. This should perhaps be changed to make sure we get good detection at the ends of pieces, and also so that dissonance detection works on short excerpts in the same way it would on entire pieces.